requirements.txt
asgiref==3.2.5
astroid==2.3.3
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect2==0.3.0
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-crispy-forms==1.9.1
django-jsonfield==1.4.0
django-markdown-deux==1.0.5
django-mathfilters==1.0.0
django-notify-x==0.1.9
django-pagedown==2.0.3
django-tinymce==3.0.2
gunicorn==20.0.4
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
markdown2==2.3.9
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.1.1
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2019.3
reportlab==3.5.42
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
stripe==2.46.0
tinycss2==1.0.2
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.12.1
xhtml2pdf==0.2.4
zope.interface==5.1.0
